货品编号->条码 货品管理功能完善 sql更新

+++ spt/src/main/java/com/hit/spt/controller/InventoryController.java
@@ -101,7 +101,7 @@ public class InventoryController {
                 JSONObject jbo = JSONObject.parseObject(inv);
                 Inventory inventory1 = new Inventory();
                 inventory1.setI_id(jbo.getInteger("i_id"));
-                inventory1.setG_id(jbo.getInteger("g_id"));
+                inventory1.setG_id(jbo.getLong("g_id"));
                 inventory1.setQuality(jbo.getString("quality"));
                 inventory1.setQuantity(jbo.getInteger("quantity"));
                 inventoryService.updateInventory(inventory1);
