+++ spt/src/main/java/com/hit/spt/pojo/OrderItem.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class OrderItem {
     private Integer oi_id;
     private Integer o_id;
-    private Integer g_id;
+    private Long g_id;
     private Integer quantity;
     private Double cost;
     private Double price;
